requirements.txt
aniso8601==8.0.0
asgiref==3.2.3
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.2
djangorestframework==3.11.0
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
Python-Rest-Framework==0.3.10
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
Werkzeug==0.16.0
